LPD-90763 Share asset entry item selector props across callers
